forgot to save again

+++ Obstacles/Camera.cs
@@ -1,4 +1,4 @@
 namespace Obstacles;
 class Camera : Obstacle {
-    public Camera(byte x, byte y) : base(x, y) {}
+    public Camera(int x, int y) : base(x, y) {}
 }
